// src/components/events/types.ts
export interface EventItem {
  collectionId: string;
  collectionName: string;
  externalId: string;
  location: string;
  id: string;
  date: Date;
  lastModified: Date;
  duration: Number;
  freeEvent: boolean;
  title: string;
  shortDescription: string,
  headerImageUrl: string;
  eventUrl: string;
}